requirements.txt
tensorflow>=2.10.0
opencv-python
numpy
matplotlib
scikit-image

streamlit
Pillow
